+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased] - ReleaseDate
 
+## [0.2.10] - 2025-12-15
+
 ### Added
 
 - Implement GetDrawColor
@@ -141,8 +143,9 @@
 - First release
 
 <!-- next-url -->
+[Unreleased]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.10...HEAD
 
-[Unreleased]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.9...HEAD
+[0.2.10]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.9...v0.2.10
 [0.2.9]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.8...v0.2.9
 [0.2.8]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.7...v0.2.8
 [0.2.7]: https://github.com/meehl/rusty-path-of-building/compare/v0.2.6...v0.2.7